+++ Alpaca/node.h
@@ -2,16 +2,17 @@
 
 #include <set>
 #include <vector>
+#include <stdint.h>
 
 struct node
 {
 	int x;
 	int y;
-	std::vector<int> distanceArray;
+	std::vector<uint16_t> distanceArray;
 
 	node(int x, int y);
 	void preCalcStart();
-	void preCalcDistance(int index, std::multiset<int>* minPathArray);
+	void preCalcDistance(int index, std::multiset<uint16_t>* minPathArray);
 	void draw(int index, int radius);
 };
